Deleting Tags

+++ application/controllers/Admin.php
@@ -25,7 +25,7 @@ class Admin extends CI_Controller {
         }
         $npk = $this->session->userdata('npk');
         $data['admin']      = $detailEmployee;
-        $data['tags']        = $this->AdminM->getTags();
+        $data['tags']       = $this->AdminM->getTags();
         $data['countAdmin'] = $this->AdminM->getAdminTotal();
         $data['employee']   = $this->OracleDBM->getAllEmp();
         $data['dept']       = $this->OracleDBM->getAllDept();
@@ -55,7 +55,7 @@ class Admin extends CI_Controller {
                 $this->AdminM->saveAdmin((string)$admin);
             }
         }
-        // redirect('Admin');
+        redirect('Admin');
 	}
 
 	public function deleteAdmin($npk)
@@ -77,7 +77,11 @@ class Admin extends CI_Controller {
     public function deleteTag($id)
 	{
 		if (!$this->isAllowed()) return redirect(site_url());
-        $this->AdminM->deleteTag($id);
+        if ($this->AdminM->checkIsTagExist($id)) {
+            $this->session->set_flashdata('error_message', 'Tidak dapat menghapus tag karena masih digunakan dalam pelatihan.');
+        } else {
+            $this->AdminM->deleteTag($id);
+        }
         redirect(site_url('Admin'));
 	}
 }
